create new components out of it
